script-gerador/requirements.txt
jinja2==3.1.3
openai==1.14.0
google-genai==0.2.0
python-dotenv==1.0.1
